Updated job/daemonset/replicationcontroller controls to delete resources using DeletePropagationForeground as propagation policy.

+++ rcc.go
@@ -116,7 +116,8 @@ func (c *RCControl) collectPods(replicationController *v1.ReplicationController)
 }
 
 func (c *RCControl) Delete(ctx context.Context, cascade bool) error {
-	c.Infof("Delete")
+	c.Infof("delete %v", formatMeta(c.replicationController.ObjectMeta))
+
 	rcs := c.Client.Core().ReplicationControllers(c.replicationController.Namespace)
 	currentRC, err := rcs.Get(c.replicationController.Name, metav1.GetOptions{})
 	if err != nil {
@@ -127,47 +128,55 @@ func (c *RCControl) Delete(ctx context.Context, cascade bool) error {
 	if err != nil {
 		return trace.Wrap(err)
 	}
-	c.Infof("deleting")
-	err = rcs.Delete(c.replicationController.Name, nil)
+	c.Info("deleting current replication controller")
+	deletePolicy := metav1.DeletePropagationForeground
+	err = rcs.Delete(c.replicationController.Name, &metav1.DeleteOptions{
+		PropagationPolicy: &deletePolicy,
+	})
 	if err != nil {
 		return ConvertError(err)
 	}
 	if !cascade {
-		c.Infof("cascade not set, returning")
+		c.Info("cascade not set, returning")
 	}
-	c.Infof("deleting pods %v", len(currentPods))
-	for _, pod := range currentPods {
-		log.Infof("deleting pod %v", pod.Name)
-		if err := pods.Delete(pod.Name, nil); err != nil {
-			return ConvertError(err)
-		}
-	}
-	return nil
+
+	err = deletePodsList(pods, currentPods, *c.Entry)
+	return trace.Wrap(err)
 }
 
 func (c *RCControl) Upsert(ctx context.Context) error {
-	c.Infof("Upsert")
+	c.Infof("upsert %v", formatMeta(c.replicationController.ObjectMeta))
+
 	rcs := c.Client.Core().ReplicationControllers(c.replicationController.Namespace)
-	c.replicationController.UID = ""
-	c.replicationController.SelfLink = ""
-	c.replicationController.ResourceVersion = ""
 	currentRC, err := rcs.Get(c.replicationController.Name, metav1.GetOptions{})
 	err = ConvertError(err)
 	if err != nil {
 		if !trace.IsNotFound(err) {
 			return trace.Wrap(err)
 		}
+		currentRC = nil
+	}
+
+	if currentRC != nil {
+		control, err := NewRCControl(RCConfig{ReplicationController: currentRC, Client: c.Client})
+		if err != nil {
+			return ConvertError(err)
+		}
+		err = control.Delete(ctx, true)
+		if err != nil {
+			return ConvertError(err)
+		}
+	}
+
+	c.replicationController.UID = ""
+	c.replicationController.SelfLink = ""
+	c.replicationController.ResourceVersion = ""
+
+	err = withExponentialBackoff(func() error {
 		_, err = rcs.Create(&c.replicationController)
 		return ConvertError(err)
-	}
-	// delete all pods and the old replication controller
-	control, err := NewRCControl(RCConfig{ReplicationController: currentRC, Client: c.Client})
-	err = control.Delete(ctx, true)
-	if err != nil {
-		return ConvertError(err)
-	}
-	_, err = rcs.Create(&c.replicationController)
-	return ConvertError(err)
+	})
+	return trace.Wrap(err)
 }
 
 func (c *RCControl) nodeSelector() labels.Selector {
